Move OptimisedExponentialMidiTokenizer to ExponentialMidiTokenizer, repair self.pad_token_id in MidiTokenizer
